v1.3.0 Major updates for ticket page
- Add camping map to ticket page for user-selection

- Add more validation rules to ticket registration process

- Minor changes to webpage layout

- Add class Visitor for manipulating visitors information with database

- Rewrite class Email to use cleaner code

- Update login page with data validation

- Add reset password function to login page

- Add PHP script to auto expire reset password tokens

+++ lib/classes/Email.php
@@ -22,37 +22,36 @@ require("../lib/dompdf/autoload.inc.php");
 
 class Email {
     private $mail;
-    public function __construct($fullName, $email, $password, $ticketId, $url, $campingSpot) {
-        try {
-            //Instantiate a mail object
-            $this->mail = new PHPMailer(true);
+    public function __construct($senderEmail, $senderName, $recipentEmail, $recipentName, $body, $subject = null, $pdf = null) {
+        //Instantiate a mail object
+        $this->mail = new PHPMailer(true);
 
-            //Set server settings
-            $this->mail->isSMTP();
-            $this->mail->Host = 'smtp.googlemail.com';
-            $this->mail->SMTPAuth = true;
-            $this->mail->Username = 'wilsonjill870@gmail.com';
-            $this->mail->Password = '21121998';
-            $this->mail->SMTPSecure = 'tls';
-            $this->mail->Port = 587;
-            $this->mail->SMTPOptions = array(
-                    'ssl' => array(
-                    'verify_peer' => FALSE,
-                    'verify_peer_name' => FALSE,
-                    'allow_self_signed' => TRUE
-                )
-            ); 
-            
-            //Set recipients information
-            $this->mail->setFrom('wilsonjill870@gmail.com', 'Sportify Team');
-            $this->mail->addAddress($email, $fullName);
+        //Set server settings
+        $this->mail->isSMTP();
+        $this->mail->Host = 'smtp.googlemail.com';
+        $this->mail->SMTPAuth = true;
+        $this->mail->Username = 'wilsonjill870@gmail.com';
+        $this->mail->Password = '21121998';
+        $this->mail->SMTPSecure = 'tls';
+        $this->mail->Port = 587;
+        $this->mail->SMTPOptions = array(
+                'ssl' => array(
+                'verify_peer' => FALSE,
+                'verify_peer_name' => FALSE,
+                'allow_self_signed' => TRUE
+            )
+        );
+        
+        //Set recipients information
+        $this->mail->setFrom($senderEmail, $senderName);
+        $this->mail->addAddress($recipentEmail, $recipentName);
 
-            //Prepare email body
-            $this->mail->isHTML(true);
-            $this->mail->Subject = 'E-ticket';
-            $body = str_replace(array("%fullName%", "%email%" , "%password%", "%ticketId%", "%url%", "%campingSpot%"), array($fullName, $email, $password, $ticketId, $url, $campingSpot), file_get_contents("../lib/templates/email.inc.html"));
-            $this->mail->Body = $body;
+        //Prepare email body
+        $this->mail->isHTML(true);
+        $this->mail->Subject = $subject;
+        $this->mail->Body = $body;
 
+        if ($pdf != null) {
             //Prepare PDF attachment
             //Set DOMPDF settings
             $options = new Options();
@@ -68,24 +67,29 @@ class Email {
             $dompdf->setHttpContext($contxt);
 
             //Create PDF attachment
-            $content = str_replace(array("%fullName%", "%email%" , "%password%", "%ticketId%", "%url%", "%campingSpot%"), array($fullName, $email, $password, $ticketId, $url, $campingSpot), file_get_contents("../lib/templates/ticket.inc.html"));
-            $dompdf->loadHtml($content);
+            $dompdf->loadHtml($pdf);
             $dompdf->render();
             $output = $dompdf->output();
             file_put_contents("../resources/tmp/ticket.pdf", $output);
 
             //Attach PDF file to email
             $this->mail->addAttachment("../resources/tmp/ticket.pdf");
-
+        }
+    }
+    
+    public function Send() {
+        try {
             //Send email
             $this->mail->send();
             
-            //Delete temp file
-            unlink("../resources/tmp/ticket.pdf");
+            //Delete temp ticket file if exists
+            if (file_exists("../resources/tmp/ticket.pdf")) {
+                unlink("../resources/tmp/ticket.pdf");
+            }
             
         } catch (Exception $e) {
-            echo 'Message could not be sent. ';
-            echo 'Mailer Error: ' . $this->mail->ErrorInfo;
+            echo("Message could not be sent.");
+            echo("Mailer Error: " . $this->mail->ErrorInfo);
         }
     }
 }
